refactors artist series entry to use flatlist
