+++ src/lib/Scenes/ArtistSeries/ArtistSeries.tsx
@@ -1,117 +1,46 @@
-import { Box, EntityHeader, Flex, Sans } from "@artsy/palette"
+import { Box, Flex, Spacer, Theme } from "@artsy/palette"
 import { ArtistSeries_artistSeries } from "__generated__/ArtistSeries_artistSeries.graphql"
-import { ArtistSeriesFollowMutation } from "__generated__/ArtistSeriesFollowMutation.graphql"
 import { ArtistSeriesQuery } from "__generated__/ArtistSeriesQuery.graphql"
-import OpaqueImageView from "lib/Components/OpaqueImageView/OpaqueImageView"
-import { ReadMore } from "lib/Components/ReadMore"
-import SwitchBoard from "lib/NativeModules/SwitchBoard"
+import { AboveTheFoldFlatList } from "lib/Components/AboveTheFoldFlatList"
 import { defaultEnvironment } from "lib/relay/createEnvironment"
+import { ArtistSeriesHeader } from "lib/Scenes/ArtistSeries/ArtistSeriesHeader"
+import { ArtistSeriesMetaFragmentContainer } from "lib/Scenes/ArtistSeries/ArtistSeriesMeta"
 import renderWithLoadProgress from "lib/utils/renderWithLoadProgress"
-import { useScreenDimensions } from "lib/utils/useScreenDimensions"
-import React, { useRef } from "react"
-import { TouchableOpacity, TouchableWithoutFeedback, View } from "react-native"
-import { commitMutation, createFragmentContainer, graphql, QueryRenderer, RelayProp } from "react-relay"
-
+import React from "react"
+import { View } from "react-native"
+import { createFragmentContainer, graphql, QueryRenderer } from "react-relay"
 interface ArtistSeriesProps {
   artistSeries: ArtistSeries_artistSeries
-  relay: RelayProp
 }
 
-type ArtistToFollowOrUnfollow = NonNullable<NonNullable<ArtistSeriesProps["artistSeries"]["artists"]>[0]>
-
-export const ArtistSeries: React.FC<ArtistSeriesProps> = ({ artistSeries, relay }) => {
-  const { width } = useScreenDimensions()
-  const isIPad = width > 700
-  const maxChars = isIPad ? 200 : 120
-  const ref = useRef<View | null>(null)
-
-  const followOrUnfollowArtist = (followArtist: ArtistToFollowOrUnfollow) => {
-    return new Promise<void>((resolve, reject) => {
-      commitMutation<ArtistSeriesFollowMutation>(relay.environment, {
-        mutation: graphql`
-          mutation ArtistSeriesFollowMutation($input: FollowArtistInput!) {
-            followArtist(input: $input) {
-              artist {
-                isFollowed
-              }
-            }
-          }
-        `,
-        variables: {
-          input: { artistID: followArtist.internalID, unfollow: followArtist.isFollowed },
-        },
-        onError: reject,
-        onCompleted: (_response, errors) => {
-          if (errors && errors.length > 0) {
-            reject(new Error(JSON.stringify(errors)))
-          } else {
-            resolve()
-          }
-        },
-      })
-    })
-  }
-
+export const ArtistSeries: React.FC<ArtistSeriesProps> = ({ artistSeries }) => {
+  const sections = ["artistSeriesMeta", "artistSeriesArtwork"]
   return (
-    <View ref={ref}>
-      <Box p={2}>
-        <Flex flexDirection="row" justifyContent="center">
-          {/* TODO: add image url */}
-          <OpaqueImageView width={180} height={180} />
-        </Flex>
-        <Sans size="8" mt={3} data-test-id="title">
-          {artistSeries.title}
-        </Sans>
-        {(artistSeries?.artists ?? []).length > 0 &&
-          artistSeries.artists?.map(
-            artist =>
-              !!artist?.name && (
-                <TouchableOpacity
-                  key={artist?.id as string}
-                  onPress={() => {
-                    SwitchBoard.presentNavigationViewController(ref.current!, `/artist/${artist.slug}`)
-                  }}
-                  style={{ marginVertical: 10 }}
-                >
-                  <EntityHeader
-                    smallVariant
-                    name={artist?.name}
-                    imageUrl={artist?.image?.url as string}
-                    FollowButton={
-                      <TouchableWithoutFeedback
-                        onPress={() => followOrUnfollowArtist(artist)}
-                        hitSlop={{ top: 20, left: 20, bottom: 20, right: 20 }}
-                      >
-                        <Sans style={{ textDecorationLine: "underline" }} size="3">
-                          {artist.isFollowed ? "Following" : "Follow"}
-                        </Sans>
-                      </TouchableWithoutFeedback>
-                    }
-                  />
-                </TouchableOpacity>
-              )
-          )}
-        <ReadMore data-test-id="description" sans content={artistSeries?.description ?? ""} maxChars={maxChars} />
+    <Theme>
+      <Box px={2}>
+        <AboveTheFoldFlatList
+          showsVerticalScrollIndicator={false}
+          initialNumToRender={3}
+          data={sections}
+          ListHeaderComponent={<ArtistSeriesHeader />}
+          ItemSeparatorComponent={() => <Spacer mb={2} />}
+          keyExtractor={(_item, index) => String(index)}
+          renderItem={({ item }): null | any => {
+            switch (item) {
+              case "artistSeriesMeta":
+                return <ArtistSeriesMetaFragmentContainer artistSeries={artistSeries} />
+            }
+          }}
+        />
       </Box>
-    </View>
+    </Theme>
   )
 }
 
 export const ArtistSeriesFragmentContainer = createFragmentContainer(ArtistSeries, {
   artistSeries: graphql`
     fragment ArtistSeries_artistSeries on ArtistSeries {
-      title
-      description
-      artists(size: 1) {
-        id
-        internalID
-        name
-        slug
-        isFollowed
-        image {
-          url
-        }
-      }
+      ...ArtistSeriesMeta_artistSeries
     }
   `,
 })
